Use float cmake result1 cursor wobble

+++ src/cmake.cpp
@@ -3259,7 +3259,7 @@ void CMenuPcs::CmakeResultDraw1()
     DrawInit();
 
     if (mode == 1) {
-        int cursorX = static_cast<int>(FLOAT_80333304) + (System.m_frameCounter & 7);
+        int cursorX = static_cast<int>(FLOAT_80333304 + static_cast<float>(System.m_frameCounter & 7));
         int cursorY = 0x70 + *reinterpret_cast<short*>(state + 0x26) * 0x28;
         DrawCursor(cursorX, cursorY, alpha);
     }
